Backend/requirements.txt
fastapi
uvicorn
python-jose
passlib[bcrypt]
groq
python-dotenv